Clear (global) environment and free up memory

Removes all objects in the global environment including hidden objects, and afterwards runs the garbage collector to free up memory.

<<<<<<< Updated upstream
clearEnv(envir = .GlobalEnv, hidden = TRUE)
=======
clearEnv(envir = .GlobalEnv, hidden = TRUE)
>>>>>>> Stashed changes
Arguments

envir
(optional) The environment to clear. default: .GlobalEnv

hidden
(optional) Include hidden objects (those starting with a dot "."). default: TRUE

Value

Prints the removed objects, their classes, and the output from garbage collector.

Examples
<<<<<<< Updated upstream

assign("test", "test", envir = .GlobalEnv)
clearEnv()

#> The following 2 objects were removed:
=======
    

assign("test", "test", envir = .GlobalEnv)
clearEnv()

#> The following 2 objects were removed:
>>>>>>> Stashed changes
#>  *object*     *class*  
#>  .Random.seed integer  
#>  test         character
#> 
#> Garbage collector:
#> 

#>           used (Mb) gc trigger (Mb) max used (Mb)
<<<<<<< Updated upstream
#> Ncells  770267 41.2    1409377 75.3   770267 41.2
#> Vcells 1411883 10.8    8388608 64.0  1411883 10.8

<<<<<<< Updated upstream ======= >>>>>>> Stashed changes