Get the column classes of a dataframe

Returns a vector with the column classes of a dataframe like object, or coerces the column classes as defined by user.

<<<<<<< Updated upstream
colClasses(df)
=======
colClasses(df)
>>>>>>> Stashed changes
Arguments

df
Dataframe like object

Value

A dataframe containing the column classes of the provided dataframe

Note

Cannot handle columns with more than one class.

Examples

df <- iris[1:10, ]
df

#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1           5.1         3.5          1.4         0.2  setosa
#> 2           4.9         3.0          1.4         0.2  setosa
#> 3           4.7         3.2          1.3         0.2  setosa
#> 4           4.6         3.1          1.5         0.2  setosa
#> 5           5.0         3.6          1.4         0.2  setosa
#> 6           5.4         3.9          1.7         0.4  setosa
#> 7           4.6         3.4          1.4         0.3  setosa
#> 8           5.0         3.4          1.5         0.2  setosa
#> 9           4.4         2.9          1.4         0.2  setosa
#> 10          4.9         3.1          1.5         0.1  setosa

colClasses(df)

#>                class
#> Sepal.Length numeric
#> Sepal.Width  numeric
#> Petal.Length numeric
#> Petal.Width  numeric
#> Species       factor

<<<<<<< Updated upstream

<<<<<<< Updated upstream ======= >>>>>>> Stashed changes