Document, check, and install a package

Everything I do before committing changes to an R package on GitHub. First restart R, then generates man pages, checks, installs, and builds package documentation website etc.

ksa_gogoawesomepkg(style = TRUE, check = TRUE, install = TRUE,
  pkgdown = TRUE, verbose = TRUE,
  num_threads = parallel::detectCores() - 2L)

Arguments

style
Style the source files found in the "R/" directory using style_file. This is performed in parallel by default on the individual files as it is much faster than simply using style_dir. Set num_threads to 1 to instead run style_dir. (Default: TRUE)

check
Check the package using check or not. (Default: TRUE)

install
Install the package (TRUE) using install or just load it in memory (FALSE) using load_all. (Default: TRUE)

pkgdown
Build package documentation website using build_site or not. (Default: TRUE)

verbose
Print status messages during the process or not. (Default: TRUE)

num_threads
The number of threads to be used if style is set to TRUE. (Default: parallel::detectCores() - 2L)

Value

Invisibly returns a list with the output of all steps run.